+++ src/framework/graph-visualiser/canvas/zoom-controls/graph-zoom-controls.component.scss
