+++ src/framework/graph-visualiser/engine/node-programs/diamond/index.ts
