+++ src/framework/graph-visualiser/engine/node-programs/diamond/shader-vert.ts
