+++ src/framework/graph-visualiser/engine/node-programs/diamond/utils.ts
