+++ src/framework/graph-visualiser/engine/node-programs/ellipse/index.ts
