+++ src/framework/graph-visualiser/engine/node-programs/ellipse/shader-vert.ts
