+++ src/framework/graph-visualiser/engine/node-programs/ellipse/utils.ts
