+++ src/framework/graph-visualiser/engine/node-programs/rounded-rect/shader-frag.ts
