+++ src/framework/graph-visualiser/engine/node-programs/rounded-rect/utils.ts
