+++ src/framework/graph-visualiser/engine/sigma-label-utils.ts
