+++ src/framework/graph-visualiser/engine/sigma-settings.ts
