+++ src/framework/graph-visualiser/style-editor/editor-tab.component.html
