+++ src/framework/graph-visualiser/style-editor/graph-styles-pane.component.scss
