+++ src/framework/graph-visualiser/style-editor/graph-styles-pane.component.ts
@@ -3,8 +3,8 @@ import { MatSelectModule } from "@angular/material/select";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatSlideToggleModule } from "@angular/material/slide-toggle";
-import { GraphStyleService } from "../../service/graph-style.service";
-import { GraphVisualiser } from "../graph-visualiser";
+import { GraphStyleService } from "../../../service/graph-style.service";
+import { GraphVisualiser } from "../engine";
 import { HighlightsTabComponent } from "./highlights-tab.component";
 import { ThemesTabComponent } from "./themes-tab.component";
 import { EditorTabComponent } from "./editor-tab.component";
