+++ src/service/query-page-state.service.ts
@@ -8,10 +8,10 @@ import { inject, Injectable } from "@angular/core";
 import { FormControl } from "@angular/forms";
 import { BehaviorSubject, combineLatest, map, Observable, pairwise, shareReplay, startWith, switchMap } from "rxjs";
 import { DriverAction } from "../concept/action";
-import { GraphVisualiser } from "../framework/graph-visualiser";
-import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/sigma-settings";
-import { newGraph, Graph } from "../framework/graph-visualiser/graph";
-import { Layouts } from "../framework/graph-visualiser/layout";
+import { GraphVisualiser } from "../framework/graph-visualiser/engine";
+import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/engine/sigma-settings";
+import { newGraph, Graph } from "../framework/graph-visualiser/engine/graph";
+import { Layouts } from "../framework/graph-visualiser/engine/layout";
 import { detectOS } from "../framework/util/os";
 import { INTERNAL_ERROR } from "../framework/util/strings";
 import { DriverState } from "./driver-state.service";
